+++ Classes/DDTTYLogger.m
@@ -1161,6 +1161,10 @@ static DDTTYLogger *sharedInstance;
     }
 }
 
+- (int)stdForLogMessage:(DDLogMessage *)logMessage {
+    return logMessage->_level == DDLogLevelError ? STDERR_FILENO : STDOUT_FILENO;
+}
+
 - (void)logMessage:(DDLogMessage *)logMessage {
     NSString *logMsg = logMessage->_message;
     BOOL isFormatted = NO;
@@ -1227,7 +1231,8 @@ static DDTTYLogger *sharedInstance;
             return;
         }
 
-        // Write the log message to STDERR
+        // Write the log message to `std`
+        int std = [self stdForLogMessage: logMessage];
 
         if (isFormatted) {
             // The log message has already been formatted.
@@ -1262,7 +1267,7 @@ static DDTTYLogger *sharedInstance;
                 v[3].iov_len = (msg[msgLen] == '\n') ? 0 : 1;
             }
 
-            writev(STDERR_FILENO, v, iovec_len);
+            writev(std, v, iovec_len);
         } else {
             // The log message is unformatted, so apply standard NSLog style formatting.
 
@@ -1357,7 +1362,7 @@ static DDTTYLogger *sharedInstance;
             v[11].iov_base = "\n";
             v[11].iov_len = (msg[msgLen] == '\n') ? 0 : 1;
 
-            writev(STDERR_FILENO, v, 13);
+            writev(std, v, 13);
         }
 
         if (!useStack) {